Moved the tests to src/test/java
