+++ src/test/java/entity/MessageUnitTest.java
@@ -1,5 +1,7 @@
 package entity;
 
+
+
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertSame;
